remove unused files and move post components to articles folder
