--- src/svgs/plus.tsx
@@ -1,12 +1,0 @@
-export const PlusSvg = () => (
-  <svg
-    width="18"
-    height="19"
-    viewBox="0 0 18 19"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <rect y="8.3335" width="18" height="2.66667" fill="#362E83" />
-    <rect y="8.3335" width="18" height="2.66667" fill="#362E83" />
-  </svg>
-);
